Actors Module

Actors.Actors — Module

Actors implements the classical Actor Model and is based on the primitives defined in ActorInterfaces.Classic. It provides:

basic primitives for creating actors, sending messages to them and changing behavior: spawn, send, become with Addr and self,

onmessage, executed by an actor on a received message,

a message protocol with predefined messages,

an API based on the protocol with primitives receive and request and further API functions become!, call, cast, exec, exit!, init!, query, term!, update!,

error handling with actor

connections: connect, disconnect, trapExit,

monitors: monitor, demonitor,

supervisors: supervisor, supervise, unsupervise, start_actor, start_task, count_children, which_children, terminate_child,

an actor registry: register, unregister, whereis, registered

and more.

The current stable, registered version is installed with

pkg> add Actors

The development version is installed with:

pkg> add "https://github.com/JuliaActors/Actors.jl"

Actors.version — Constant

Gives the package version.

julia> using Actors

julia> Actors.version
v"0.2.4"