Actor Registry

Actors can be registered with Symbols to a registry. API functions on actors can then be called with their registered names.

Actors.register — Function

register(name::Symbol, lk::Link)

Register the actor lk with name. Returns true if the registration succeeds, false if name is already in use.

Actors.unregister — Function

unregister(name::Symbol)

Remove any registrations associated with name.

Actors.whereis — Function

whereis(name::Symbol)

Find out whether name is registered. Return the actor link lk or missing if not found.

Actors.registered — Function

registered()

Return an Array of all registered actors in the system.

The registry works transparently over distributed worker processes such that local links are transformed to remote links when shared between workers.