Actor API

Installation

Actors.Actors — Module

Actors

A Julia library implementing the classical Actor Model.

The current stable, registered version is installed with

pkg> add Actors

The development version is installed with:

pkg> add("https://github.com/JuliaActors/Actors.jl")

Actors.version — Constant

Gives the package version.

Types

Actors.Msg — Type

Abstract type for messages to actors.

Actors.Request — Type

Request(x, from::Link)

A generic Msg for user requests.

Actors.Response — Type

Response(y, from::Link=self())

A Msg representing a response to requests.

Fields

y: response content,

from::Link: sender link.

Actors.Link — Type

Link{C}(chn::C, pid::Int, type::Symbol)

A mailbox for communicating with actors. A concrete type of this must be returned by an actor on creation with spawn.

Fields/Parameters

chn::C: C can be any type and characterizes the interface to an actor,

pid::Int: the pid of the actor,

type::Symbol: an arbitrary symbol characterizing the actor.

Actors.Func — Type

Func(f, args...; kwargs...)

A structure for passing a function f and its arguments to an actor.

Actors._ACT — Type

_ACT()

Internal actor status variable.

Fields

bhv::Func : the behavior function and its internal arguments,

init::Func: the init function and its arguments,

term::Func: the terminate function and its arguments,

self::Link: the actor's (local or remote) self,

name::Symbol: the actor's registered name.

res::Any: the result of the last behavior execution,

usr::Any: user variable for plugging in something.

see also: Func, Link

Primitives

Actors.spawn — Function

spawn(bhv::Func; pid=myid(), thrd=false, sticky=false, taskref=nothing)
spawn(m::Val(:Actors), args...; kwargs...)
spawn(m::Module, args...; kwargs...)

Start a function bhv as an actor and return a Link to it.

Parameters

bhv::Func: behavior function,

pid=myid(): pid of worker process the actor should be started on,

thrd=false: thread number the actor should be started on or false,

sticky=false: if true the actor is started on the current thread,

taskref=nothing: if a Ref{Task}() is given here, it gets the started Task,

m::Module: the Module implementing spawn.

Actors.send! — Function

send!(lk::Link, msg)

Send a message to an actor.

Actors.become! — Function

become!(lk::Link, bhv::Func)

Tell an actor lk to assume the behavior function bhv.

Actors.become — Function

become(bhv::Function, args...; kwargs...)

Cause your actor to take on a new behavior. This can only be called from inside an actor/behavior.

Arguments

bhv::Function: function implementing the new behavior,

args...: arguments to bhv (without msg),

kwargs...: keyword arguments to bhv.

Actors.self — Function

self()

Get the Link of your actor.

Actors.onmessage — Function

onmessage(A::_ACT, msg)

An actor executes this function when a message arrives. An application can extend this by further methods and use it to plugin the Actors.jl API.

User API

Actors.receive! — Function

receive!(lk; timeout=5.0)
receive!(lk, from; timeout=5.0)
receive!(lk, M; timeout=5.0)
receive!(lk, M, from; timeout=5.0)

Receive a message over a link lk.

If M or from are provided, receive! returns only a matching message. Other messages in lk are restored to it in their previous order.

Parameters

lk::Link: local or remote link over which the message is received,

M::Type{<:Msg}: Msg type,

from::Link: local or remote link of sender. If from is provided, only messages with a from field can be matched.

timeout::Real=5.0: maximum waiting time in seconds.

If timeout==0, lk is scanned only for existing messages.

Set timeout=Inf if you don't want to timeout.

Returns

received message or Timeout().

Actors.request! — Function

request!(lk::Link, msg::Msg; full=false, timeout::Real=5.0)
request!(lk::Link, M::Type{<:Msg}, args...; kwargs...)

Send a message to an actor, block, receive and return the result.

Arguments

lk::Link: actor link, or name::Symbol (if registered),

msg::Msg: a message,

Msg::Type{<:Msg}: a message type,

args...: optional arguments to Msg,

full: if true return the full Response message.

timeout::Real=5.0: timeout in seconds after which a Timeout is returned,

kwargs...: full or timeout.

Internal Messages

Actors.Become — Type

Become(x::Func)

An asynchronous Msg to an actor to change its behavior.

Actors.Diag — Type

Diag(from::Link)

A synchronous Msg to an actor to send a Response message with its internal _ACT variable to from.

Actors.Exit — Type

Exit(reason, stack)

A Msg causing an actor to stop with an exit code. If present, it calls its term! function with code as last argument.

Actors.Update — Type

Update(s::Symbol, x)

An asynchronous Msg to an actor to update its internal state s to x.

s::Symbol can be one of :arg, :self, :usr.

Actors.Timeout — Type

Timeout()

A return value to signal that a timeout has occurred.

Diagnosis