Actor API

Installation

Actors.Actors — Module

Actors

A Julia library implementing the classical Actor Model.

The current stable, registered version is installed with

pkg> add Actors

The development version is installed with:

pkg> add "https://github.com/JuliaActors/Actors.jl"

Actors.version — Constant

Gives the package version.

julia> using Actors

julia> Actors.version
v"0.1.3"

Types

The following types are needed for using and extending Actors:

Actors.Msg — Type

Abstract type for messages to actors.

Actors.Request — Type

Request(x, from::Link)

A generic Msg for user requests.

Actors.Response — Type

Response(y, from::Link=self())

A Msg representing a response to requests.

Fields

y: response content,

from::Link: sender link.

Actors.Link — Type

Link{C}(chn::C, pid::Int, type::Symbol)

A mailbox for communicating with actors. A concrete type of this must be returned by an actor on creation with spawn.

Fields/Parameters

chn::C: C can be any type and characterizes the interface to an actor,

pid::Int: the pid of the actor,

mode::Symbol: a symbol characterizing the actor mode.

Actors.Func — Type

Func(func, a...; kw...)(c...)

A callable struct to represent actor behavior. It is executed with parameters from the incoming communication.

Parameters

f: a callable object (function, functor ...),

a...: acquaintance parameters to f. Those are stored,

kw...: keyword arguments,

c...: parameters from the incoming communication.

Actors._ACT — Type

_ACT{T}

Internal actor status variable.

Fields

mode::Symbol: the actor mode,

bhv::T : the behavior function and its internal arguments,

init::T: the init function and its arguments,

term::T: the terminate function and its arguments,

self::Addr: the actor's (local or remote) self,

name::Symbol: the actor's registered name.

res::Any: the result of the last behavior execution,

sta::Any: a variable for representing state,

usr::Any: user variable for plugging in something.

see also: Func, Link

Starting Actors, creating links

Actors.jl doesn't export its functions to start actors and to create links. Thus other libraries building on it can implement their own actors and links.

To use Actors's actors and links you import them explicitly:

using Actors
import Actors: spawn, newLink

Then you can create them with the following functions:

Actors.spawn — Function

spawn(bhv::Func; pid=myid(), thrd=false, sticky=false, taskref=nothing, mode=:default)
spawn(m::Val(:Actors), args...; kwargs...)
spawn(m::Module, args...; kwargs...)

Start a function bhv as an actor and return a Link to it.

Parameters

bhv::Func: behavior function,

pid=myid(): pid of worker process the actor should be started on,

thrd=false: thread number the actor should be started on or false,

sticky=false: if true the actor is started on the current thread,

taskref=nothing: if a Ref{Task}() is given here, it gets the started Task,

mode=:default: mode, the actor should operate in.

Actors.newLink — Function

newLink(size=32; remote=false, pid=myid(), mode=nothing)

Create a local Link with a buffered Channel size ≥ 1.

Parameters

size=32: the size of the channel buffer,

remote=false: should a remote link be created,

pid=myid(): optional pid of the remote worker,

mode=nothing: if mode==nothing the mode is automatically set to :local or :remote.

Actor Primitives

The following primitives characterize actors in the classical Actor Model:

Actors.send — Function

send(lk::Link, msg)

Send a message to an actor.

Actors.become! — Function

become!(lk::Link, bhv::Func)
become!(lk::Link, func, args1...; kwargs...)
become!(name::Symbol, ....)

Cause an actor to change behavior.

Arguments

actor lk::Link (or name::Symbol if registered),

bhv: Func implementing the new behavior,

func: callable object,

args1...: (partial) arguments to func,

kwargs...: keyword arguments to func.

Actors.become — Function

become(bhv::Func)
become(func, args...; kwargs...)

Cause your actor to take on a new behavior. This can only be called from inside an actor/behavior.

Arguments

bhv::Func: Func implementing the new behavior,

func: callable object,

args1...: (partial) arguments to func,

kwargs...: keyword arguments to func.

Actors.self — Function

self()

Get the Link of your actor.

Actors.stop — Function

stop(reason::Symbol)

Cause your actor to stop with a reason.

Actors.onmessage — Function

onmessage(A::_ACT, msg)

An actor executes this function when a message arrives. An application can extend this by further methods and must use it to plugin the Actors.jl API.

API Primitives

To receive messages from actors the following two functions for synchronous and asynchronous communication are useful:

Actors.receive — Function

receive(lk; timeout=5.0)
receive(lk, from; timeout=5.0)
receive(lk, M; timeout=5.0)
receive(lk, M, from; timeout=5.0)

Receive a message over a link lk.

If M or from are provided, receive returns only a matching message. Other messages in lk are restored to it in their previous order.

Parameters

lk::Link: local or remote link over which the message is received,

M::Type{<:Msg}: Msg type,

from::Link: local or remote link of sender. If from is provided, only messages with a from field can be matched.

timeout::Real=5.0: maximum waiting time in seconds.

If timeout==0, lk is scanned only for existing messages.

Set timeout=Inf if you don't want to timeout.

Returns

received message or Timeout().

Actors.request — Function

request(lk::Link, msg::Msg; full=false, timeout::Real=5.0)
request(lk::Link, M::Type{<:Msg}, args...; kwargs...)

Send a message to an actor, block, receive and return the result.

Arguments

lk::Link: actor link, or name::Symbol (if registered),

msg::Msg: a message,

Msg::Type{<:Msg}: a message type,

args...: optional arguments to Msg,

full: if true return the full Response message.

timeout::Real=5.0: timeout in seconds after which a Timeout is returned,

kwargs...: full or timeout.

User API

Actors has a user interface allowing you to control actors:

Actors.call! — Function

call!(lk::Link, [from::Link,] args2...; timeout::Real=5.0)
call!(name::Symbol, ....)

Call an actor to execute its behavior and to send a Response with the result.

Arguments

actor lk::Link (or name::Symbol if registered),

from::Link: sender link,

args2...: remaining arguments to the actor.

timeout::Real=5.0: timeout in seconds.

Note: If from is omitted, call! blocks and returns the result

Actors.cast! — Function

cast!(lk::Link, args2...)
cast!(name::Symbol, args2...)

Cast args2... to the actor lk (or name if registered) to execute its behavior with args2... without sending a response.

Note: you can prompt the returned value with query.

Actors.exec! — Function

exec!(lk::Link, from::Link, func, args...; kwargs...)
exec!(lk::Link, from::Link, f::Func)
exec!(lk::Link, f::Func; timeout::Real=5.0)
exec!(name::Symbol, ....)

Ask an actor lk (or name if registered) to execute an arbitrary function and to send the returned value as Response.

Arguments

actor lk::Link or name::Symbol if registered,

from::Link: the link a Response should be sent to.

func: a callable object,

args...; kwargs...: arguments and keyword arguments to it,

fu::Func: a Func with a callable object and its arguments,

timeout::Real=5.0: timeout in seconds. Set timeout=Inf if you don't want to timeout.

Note: If from is ommitted, exec! blocks, waits and returns the result (with a timeout).

Actors.exit! — Function

exit!(lk::Link, reason=:ok)
exit!(name::Symbol, ....)

Tell an actor lk (or name if registered) to exit. If it has a term function, it calls it with reason as last argument.

This behavior is not yet fully implemented!

It is needed for supervision.

Actors.query — Function

query(lk::Link, [from::Link,] s::Symbol; timeout::Real=5.0)
query(name::Symbol, ....)

Query an actor about an internal state variable s.

Parameters

actor lk::Link or name::Symbol if registered,

from::Link: sender link,

s::Symbol one of :mode,:bhv,:res,:sta,:usr.

timeout::Real=5.0:

Note: If from is omitted, query blocks and returns the response. In that case there is a timeout.

Examples

Actors.update! — Function

update!(lk::Link, x; s::Symbol=:sta)
update!(lk::Link, arg::Args)
update!(name::Symbol, ....)

Update an actor's internal state s with args....

Arguments

actor lk::Link or name::Symbol if registered,

x: value/variable to update the choosen state with,

arg::Args: arguments to update,

s::Symbol: one of :arg,:mode,:name,:self,:sta,:usr.

Note: If you want to update the stored arguments to the behavior function with s=:arg, you must pass an Args to arg. If Args has keyword arguments, they are merged with existing keyword arguments to the behavior function.

Example

The following is needed for updating arguments:

Actors.Args — Type

Args(args...; kwargs...)

A structure for updating arguments to an actor's behavior.

Actor Tasks

Actors.ATask — Type

The actor task type.

Actors.async — Function

async(func::Func; pid=myid(), thrd=false, sticky=false, taskref=nothing)

Start a task to execute func and return an ATask variable.

An actor task sends its result to the back link of the ATask variable and exits immediately.

Parameters

func::Func:

pid=myid():

thrd=false:

sticky=false:

taskref=nothing:

Actors.await — Function

await(t::ATask; timeout::Real=5.0)

Await a task reply and return it.

Actor Registry

Actors can be registered with Symbols to a registry. API functions on actors can then be called with their registered names.

Actors.register — Function

register(name::Symbol, lk::Link)

Register the actor lk with name. Returns true if the registration succeeds, false if name is already in use.

Actors.unregister — Function

unregister(name::Symbol)

Remove any registrations associated with name.

Actors.whereis — Function

whereis(name::Symbol)

Find out whether name is registered. Return the actor link lk or missing if not found.

Actors.registered — Function

registered()

Return an Array of all registered actors in the system.

The registry works transparently over distributed worker processes such that local links are transformed to remote links when shared between workers.

Actor Supervision

This is not yet implemented.

Actors.init! — Function

init!(lk::Link, func, args...; kwargs...)
init!(name::Symbol, ....)

Tell an actor lk to save the func with the given arguments as an init function and to execute it.

The init function will be called at actor restart.

This behavior is not yet implemented!

It is needed for supervision.

Actors.term! — Function

term!(lk::Link, func, args...; kwargs...)
term!(name::Symbol, ....)

Tell an actor lk (or name::Symbol if registered) to execute func with the given partial arguments and an exit reason when it terminates.

The exit reason is added by the actor to args1... when it exits.

This behavior is not yet implemented!

It is needed for supervision.

Diagnosis