Actor API

Actors.Actors — Module

Actors implements the classical Actor Model and is based on the primitives defined in ActorInterfaces.Classic. It provides:

basic primitives for creating actors, sending messages to them and changing behavior: spawn, send, become with Addr and self,

onmessage, executed by an actor on a received message,

a message protocol with predefined messages,

an API based on the protocol with primitives receive and request and further API functions become!, call, cast, exec, exit!, init!, query, term!, update!,

error handling with actor

connections: connect, disconnect, trapExit,

monitors: monitor, demonitor,

supervisors: supervisor, supervise, unsupervise, start_actor, start_task, count_children, which_children, delete_child, terminate_child,

an actor registry: register, unregister, whereis, registered

and more.

The current stable, registered version is installed with

pkg> add Actors

The development version is installed with:

pkg> add "https://github.com/JuliaActors/Actors.jl"

Actors.version — Constant

Gives the package version.

julia> using Actors

julia> Actors.version
v"0.2.4"

Basic Types

The following types are needed for using and extending Actors:

Actors.Msg — Type

Abstract type for messages to actors.

Actors.Request — Type

Request(x, from::Link)

A generic Msg for user requests.

Actors.Response — Type

Response(y, from::Link=self())

A Msg representing a response to requests.

Fields

y: response content,

from::Link: sender link.

Actors.Link — Type

Link{C} <: ActorInterfaces.Classic.Addr
Link(chn::C, pid::Int, type::Symbol) where C

A mailbox for communicating with actors. A concrete type of this must be returned by an actor on creation with spawn.

Fields/Parameters

chn::C: C can be any type and characterizes the interface to an actor,

pid::Int: the pid (worker process identifier) of the actor,

mode::Symbol: a symbol characterizing the actor mode.

Actors.Bhv — Type

Bhv(func, a...; kw...)(c...)

A callable struct to represent actor behavior. It is executed with parameters from the incoming communication.

Parameters

f: a callable object,

a...: stored acquaintance parameters to f,

kw...: stored keyword arguments,

c...: parameters from the incoming communication.

You can create your own message types with

Actors.@msg — Macro

@msg [Msg] A B C

Define empty structs as message types. Msg is an existing abstract datatype.

To call @msg Msg A B C is equivalent to

struct A <: Msg end
struct B <: Msg end
struct C <: Msg end

To call @msg D E F is equivalent to

struct D end
struct E end
struct F end

Starting Actors, creating links

Actors.jl doesn't export its functions to start actors and to create links. Thus other libraries building on it can implement their own actors and links.

To use Actors's actors and links you import them explicitly:

using Actors
import Actors: spawn, newLink

Then you can create them with the following functions:

ActorInterfaces.Classic.spawn — Function

spawn(f, args...; 
      pid=myid(), thrd=false, sticky=false, 
      taskref=nothing, remote=false, mode=:default)

Create an actor with a behavior f(args...) and return a Link to it.

Parameters

f: callable object (function, closure or functor) to execute when a message arrives,

args...: (partial) arguments to it,

pid=myid(): pid of worker process the actor should be started on,

thrd=false: thread number the actor should be started on or false,

sticky=false: if true the actor is started on the current thread,

taskref=nothing: if a Ref{Task}() is given here, it gets the started Task,

remote=false: if true, a remote channel is created,

mode=:default: mode, the actor should operate in.

Note: If you need keyword arguments kwargs... to f, you can do spawn(Bhv(f, args...; kwargs...)).

Actors.newLink — Function

newLink(size=32; remote=false, pid=myid(), mode=nothing)

Create a local Link with a buffered Channel size ≥ 1.

Parameters

size=32: the size of the channel buffer,

remote=false: should a remote link be created,

pid=myid(): optional pid of the remote worker,

mode=nothing: if mode==nothing the mode is automatically set to :local or :remote.

Actor Primitives

The following primitives characterize actors in the classical Actor Model:

ActorInterfaces.Classic.send — Function

send(lk::Link, msg...)

Send a message to an actor. msg... are communication parameters to the actor's behavior function.

Actors.become! — Function

become!(lk::Link, func, args1...; kwargs...)
become!(name::Symbol, ....)

Cause an actor to change behavior.

Arguments

actor lk::Link (or name::Symbol if registered),

func: a callable object,

args1...: (partial) arguments to func,

kwargs...: keyword arguments to func.

ActorInterfaces.Classic.become — Function

become(func, args...; kwargs...)

Cause your actor to take on a new behavior. This can only be called from inside an actor/behavior.

Arguments

func: a callable object,

args...: (partial) arguments to func,

kwargs...: keyword arguments to func.

ActorInterfaces.Classic.self — Function

self()

Get the Link of your actor.

Actors.stop — Function

stop(reason=:normal)

Cause your actor to stop with a reason.

ActorInterfaces.Classic.onmessage — Function

onmessage(bhv, msg)

Default behavior function to execute the current actor behavior bhv with the message msg. The actor calls bhv(msg) when a message arrives.

Parameters

bhv: excutable object (closure or functor) taking parameters msg,

msg: message parameters to bhv.

onmessage(A::_ACT, msg)
onmessage(A::_ACT, mode, msg)

An actor executes this function when a message arrives.

Actor libraries or applications can use this to

plugin the Actors.jl API (first form) or

extend it to other protocols by using the 2nd form.

API Primitives

To receive messages from actors the following two functions for synchronous and asynchronous communication can be used:

Actors.receive — Function

receive(lk; timeout=5.0)
receive(lk, from; timeout=5.0)
receive(lk, M; timeout=5.0)
receive(lk, M, from; timeout=5.0)

Receive a message over a link lk.

If M or from are provided, receive returns only a matching message. Other messages in lk are restored to it in their previous order.

Parameters

lk::Link: local or remote link over which the message is received,

M::Type{<:Msg}: Msg type,

from::Link: local or remote link of sender. If from is provided, only messages with a from field can be matched.

timeout::Real=5.0: maximum waiting time in seconds.

If timeout==0, lk is scanned only for existing messages.

Set timeout=Inf if you don't want to timeout.

Returns

received message or Timeout().

Actors.request — Function

request(lk::Link, msg::Msg; full=false, timeout::Real=5.0)
request(lk::Link, M::Type{<:Msg}, args...; kwargs...)

Send a message to an actor, block, receive and return the result.

Arguments

lk::Link: actor link, or name::Symbol (if registered),

msg::Msg: a message,

Msg::Type{<:Msg}: a message type,

args...: optional arguments to Msg,

full: if true return the full Response message.

timeout::Real=5.0: timeout in seconds after which a Timeout is returned,

kwargs...: full or timeout.

User API

The user API allows you to communicate with actors using the Actors protocol:

Actors.call — Function

call(lk::Link, [from::Link,] args2...; timeout::Real=5.0)
call(name::Symbol, ....)

Call an actor to execute its behavior and to send a Response with the result.

Arguments

actor lk::Link (or name::Symbol if registered),

from::Link: sender link,

args2...: remaining arguments to the actor.

timeout::Real=5.0: timeout in seconds.

Note: If from is omitted, call blocks and returns the result

Actors.cast — Function

cast(lk::Link, args2...)
cast(name::Symbol, args2...)

Cast args2... to the actor lk (or name if registered) to execute its behavior with args2... without sending a response.

Note: you can prompt the returned value with query.

Actors.exec — Function

exec(lk::Link, from::Link, f, args...; kwargs...)
exec(lk::Link, f, args...; timeout::Real=5.0, kwargs...)
exec(name::Symbol, ....)

Ask an actor lk (or name if registered) to execute an arbitrary function and to send the returned value as Response.

Arguments

actor lk::Link or name::Symbol if registered,

from::Link: the link a Response should be sent to.

f: a callable object,

args...; kwargs...: arguments and keyword arguments to it,

timeout::Real=5.0: timeout in seconds. Set timeout=Inf if you don't want to timeout.

Note: If from is ommitted, exec blocks, waits and returns the result (with a timeout).

Actors.exit! — Function

exit!(lk::Link, reason=:normal)
exit!(name::Symbol, ....)

Tell an actor lk (or name if registered) to stop. If it has a term function, it calls that with reason as last argument.

Actors.info — Function

info(lk::Link)

Return the state of an actor associated with lk:

Actors.Info if it is runnable,

:done if it has finished,

else return the failed task.

Actors.query — Function

query(lk::Link, [from::Link,] s::Symbol; timeout::Real=5.0)
query(name::Symbol, ....)

Query an actor about an internal state variable s.

Parameters

actor lk::Link or name::Symbol if registered,

from::Link: sender link,

s::Symbol one of :mode,:bhv,:res,:sta,:usr.

timeout::Real=5.0:

Note: If from is omitted, query blocks and returns the response. In that case there is a timeout.

Examples

julia> f(x, y; u=0, v=0) = x+y+u+v  # implement a behavior
f (generic function with 1 method)

julia> fact = spawn(Bhv(f, 1))     # start an actor with it
Link{Channel{Any}}(Channel{Any}(sz_max:32,sz_curr:0), 1, :default)

julia> query(fact, :mode)           # query the mode
:default

julia> cast(fact, 1)                # cast a 2nd argument to it
Actors.Cast((1,))

julia> query(fact, :res)            # query the result
2

julia> query(fact, :sta)            # query the state

julia> query(fact, :bhv)            # query the behavior
Bhv(f, (1,), Base.Iterators.Pairs{Union{},Union{},Tuple{},NamedTuple{(),Tuple{}}}(), Actors.var"#2#4"{Base.Iterators.Pairs{Union{},Union{},Tuple{},NamedTuple{(),Tuple{}}},typeof(f),Tuple{Int64}}(Base.Iterators.Pairs{Union{},Union{},Tuple{},NamedTuple{(),Tuple{}}}(), f, (1,)))

Actors.update! — Function

update!(lk::Link, x; s::Symbol=:sta)
update!(lk::Link, arg::Args)
update!(name::Symbol, ....)

Update an actor's internal state s with args....

Arguments

actor lk::Link or name::Symbol if registered,

x: value/variable to update the choosen state with,

arg::Args: arguments to update,

s::Symbol: one of :arg, :mode, :name, :self, :sta, :usr.

Note: If you want to update the stored arguments to the behavior function with s=:arg, you must pass an Args to arg. If Args has keyword arguments, they are merged with existing keyword arguments to the behavior function.

Example

julia> update!(fact, 5)       # update the state variable
Actors.Update(:sta, 5)

julia> query(fact, :sta)      # query it
5

julia> update!(fact, Args(0, u=5, v=5));  # update arguments to the behavior 

julia> call(fact, 0)          # call the actor with 0
10

The following is needed for updating arguments:

Actors.Args — Type

Args(args...; kwargs...)

A structure for updating arguments to an actor's behavior.

Actor Registry

Actors can be registered with Symbols to a registry. API functions on actors can then be called with their registered names.

Actors.register — Function

register(name::Symbol, lk::Link)

Register the actor lk with name. Returns true if the registration succeeds, false if name is already in use.

Actors.unregister — Function

unregister(name::Symbol)

Remove any registrations associated with name.

Actors.whereis — Function

whereis(name::Symbol)

Find out whether name is registered. Return the actor link lk or missing if not found.

Actors.registered — Function

registered()

Return an Array of all registered actors in the system.

The registry works transparently over distributed worker processes such that local links are transformed to remote links when shared between workers.

Error Handling

Actors.init! — Function

init!(lk::Link, f, args...; kwargs...)
init!(name::Symbol, ....)

Tell an actor lk to save the callable object f with the given arguments as an init object in its _ACT variable. The init object will be called by a supervisor at actor restart.

Arguments

actor lk::Link or name::Symbol if registered,

f: callable object,

args...: arguments to f,

kwargs...: keyword arguments to f.

Actors.term! — Function

term!(lk::Link, f, args1...; kwargs...)
term!(name::Symbol, ....)

Tell an actor lk (or name::Symbol if registered) to execute f with the given partial arguments and an exit reason when it terminates.

The exit reason is added by the actor to args1... when it exits.

Connections

Connected actors send each other Exit signals and terminate together unless they are made :sticky with trapExit:

Actors.connect — Function

connect(lk::Link)

Create a connection between the calling actor and the actor represented by lk.

Connected actors will send each other Exit signals. A connected actor will exit with the signaled reason unless it is :normal.

Note:

An actor can be made :sticky with trapExit and then will not exit.

If this is called from the Main scope, lk is

connected to the Actors._ROOT actor.

Actors.disconnect — Function

disconnect(lk::Link)

Remove the connection between the calling actor and the actor represented by lk.

Note: If this is called from the Main scope, lk is disconnected from the _ROOT actor.

Actors.trapExit — Function

trapExit(lk::Link=self(), mode=:sticky)

Change the mode of an actor.

A :sticky actor does not exit if it receives an Exit signal from a connected actor and does not propagate it further. Instead it reports the failure and saves a link to the failed actor.

See diag for getting links to failed actors from a :sticky actor.

Monitors

Actors can warn or take actions if their monitored actors or tasks terminate.

Actors.monitor — Function

monitor(lk::Link, onsignal...)
monitor(t::Task, onsignal...; timeout::Real=5.0, pollint::Real=0.1)

Start monitoring the actor represented by lk or the task t and execute onsignal... if it sends Down or if it fails.

Parameters

onsignal...: action to take on Down signal:

if empty, it gives a warning;

if it is one argument f, it executes with f(msg.reason);

if f, args..., it gets executed with f(args..., msg.reason).

timeout::Real=5.0: how many seconds should a task be monitored? After that a Down with reason :timed_out is sent.

pollint::Real=0.1: polling interval in seconds for task monitoring.

Actors.demonitor — Function

demonitor(lk::Link)

Remove the monitoring for the given link lk.

Supervisors

A supervisor actor can restart its children if they should terminate.

Actors.supervisor — Function

supervisor( strategy=:one_for_one, 
            max_restarts::Int=3, 
            max_seconds::Real=5; 
            name=nothing, kwargs...)

Start a supervisor actor with an empty child list and return a link to it.

Parameters

strategy=:one_for_one: supervision strategy, can be either :one_for_one, :one_for_all or :rest_for_one,

max_restarts::Int=3: maximum number of restarts allowed in a time frame,

max_seconds::Real=5: time frame in which max_restarts applies,

name=nothing: name (Symbol) under which it should be registered,

kwargs...: keyword arguments to spawn.

Note

See the manual chapter on error handling for explanations of strategy.

Actors.supervise — Function

supervise(sv::Link, cb=nothing, restart::Symbol=:transient)

Tell a supervisor sv to supervise the calling actor.

Parameters

sv::Link: link to a supervisor,

cb=nothing: callback (a callable object), takes the previous actor behavior as argument and must return a Link to a new actor; if nothing, the actor gets restarted with its init! callback or its previous behavior.

restart::Symbol=:transient: restart option, one of :permanent, :temporary, :transient,

Note

See the manual chapter on error handling for explanations of restart option.

Actors.unsupervise — Function

unsupervise(sv::Link)

Tell a supervisor sv to delete the calling actor from the childs list.

Actors.set_strategy — Function

set_strategy(sv::Link, strategy::Symbol)

Tell a supervisor sv to change its restart strategy.

Actors.start_actor — Function

start_actor(start, sv::Link, restart::Symbol, cb=nothing; kwargs...)

Tell a supervisor sv to start an actor, to add it to its childs list and to return a link to it.

Parameters

start: start behavior of the child, a callable object,

sv::Link: link to a started supervisor,

cb=nothing: callback (a callable object, gets the last actor behavior as argument and must return a Link); if nothing, the actor gets restarted with its init! callback or with its last behavior,

restart::Symbol=:transient: restart option, one of :permanent, :temporary, :transient,

kwargs...: keyword arguments to spawn.

Note

See the manual chapter on error handling for explanations of restart option.

Actors.start_task — Function

start_task(start, sv::Link, cb=nothing; 
           timeout::Real=5.0, pollint::Real=0.1)

Spawn a task, tell the supervisor sv to supervise it (with restart strategy :transient) and return a reference to it.

Parameters

start: must be callable with no arguments,

sv::Link: link to a started supervisor,

cb=nothing: callback for restart (a callable object, must return a Task); if cb=nothing, the task gets restarted with its start function,

timeout::Real=5.0: how long [seconds] should the task be supervised,

pollint::Real=0.1: polling interval [seconds].

Actors.count_children — Function

count_children(sv::Link)

Return a named tuple containing children counts for the given supervisor sv.

Actors.which_children — Function

which_children(sv::Link, info=false)

Tell a supervisor sv to return its childs list. If info=true it returns a list of named tuples with more child information.

Actors.delete_child — Function

delete_child(sv::Link, child)

Tell a supervisor sv to delete child (a Link or Task) from the childs list.

Actors.terminate_child — Function

terminate_child(sv::Link, child::Link)

Tell a supervisor sv to remove a child c from its childs and to terminate it with reason :shutdown.

Checkpointing

A checkpointing actor can take user-defined checkpoints from current computations and restore them on demand.

Actors.checkpointing — Function

checkpointing(level=1, filename::AbstractString=""; kwargs...)

Start a checkpointing actor and return a Link to it.

Arguments

filename::AbstractString: filename where the actor should save its checkpointing buffer,

on::Bool=false: should checkpoints automatically be updated and saved,

update::Int=10: update interval in seconds ≥ 1,

save::Int=60: saving interval in seconds ≥ 1,

kwargs...: keyword arguments to spawn.

Note

Updating or saving is only done automatically if their intervals are ≥ 1 seconds.

Actors.checkpoint — Function

checkpoint(cp::Link, key::Symbol, args...)

Tell a checkpointing actor to take a checkpoint.

Arguments

cp::Link: Link to the checkpointing actor,

key::Symbol: key for the checkpoint,

args...: variables to save in the checkpoint,

level::Int=1: checkpoint level.

Actors.restore — Function

restore(cp::Link; key::Symbol)

Tell a checkpointing actor to restore the last taken checkpoint.

Arguments

cp::Link: Link to the checkpointing actor,

key::Symbol: checkpoint key,

level::Int=1: checkpoint level to restore from.

Actors.register_checkpoint — Function

Register a checkpointing actor to a higher level actor.

Actors.@chkey — Macro

@chkey a b c 123

Build a checkpointing key from the surrounding module and function names and the given arguments. This is intended for easy construction of checkpointing keys.

Example

module MyModule

using Actors

function mykey()
    a = 1
    b = 2
    c = 3
    # do something
    return @chkey a b c 123
end

export mykey
end # MyModule

julia> using .MyModule

julia> mykey()
:MyModule_mykey_a_b_c_123

Actors.set_interval — Function

set_interval(cp::Link; kwargs...)

Set the checkpointing intervals of a checkpointing actor cp.

Arguments

cp::Link: Link to a checkpointing actor,

kwargs...: allowed keyword arguments:

on:Bool: should checkpoints automatically be updated and saved,

update::Int: update interval in seconds ≥ 1,

save::Int: saving interval in seconds ≥ 1.

Actors.get_interval — Function

Get the checkpointing intervals of a checkpointing actor.

Actors.start_checkpointing — Function

Start periodic checkpointing.

Actors.stop_checkpointing — Function

Stop periodic checkpointing.

Actors.get_checkpoints — Function

Tell the checkpointing actor cp to return its stored checkpoint data.

Actors.save_checkpoints — Function

Tell a checkpointing actor cp to save its stored checkpoint data.

Actors.load_checkpoints — Function

Tell a checkpointing actor cp to load checkpointing data from a file

Utilities

Actors.tid — Function

tid(n::Int)

Return a threadid where an n-th calculation could run.

This is for emulating a @threads for loop with actors.

Diagnosis

Actors.diag — Function

diag(lk::Link, check::Symbol=:state)
diag(name::Symbol, ....)

Diagnose an actor, get a state or stacktrace.

Arguments

lk::Link: actor link,

check::Symbol: requested information,

- `:state`: returns `:ok` if the actor is running, 
- `:task`: returns the current actor task,
- `:tid`: current actor task encoded as a proquint string,
- `:pid`: process identifier number,
- `:act`: actor `_ACT` variable,
- `:info`: actor [`Info`](@ref),
- `:err`: error log (only monitors or supervisors).

This is for diagnosis only!

Modifying an actor's state can cause a race condition.